Fix for Font Awesome CDN's auto accessibility
When Font Awesome auto-generates <span class="sr-only"></span> elements
for the social icons in the footer, the nth-child CSS selectors fail to
pick up the anchor elements as intended (they skip a number due to the
wedged <span>).
This commit fixes this behaviour by using the nth-of-type CSS selector
instead.